flask

+++ mi-primera-api/routes/tarea_routes.py
@@ -25,11 +25,17 @@ def obtener_tareas():
 
 # Busca la tarea por id
 @tareas_bp.route('/<int:id>')
+@jwt_required()
 def obtener_tarea(id):
     try:
+        usuario_id = int(get_jwt_identity())
         tarea = Tarea.query.get(id)
         if tarea is None:
             return jsonify({'ok': False, 'message': 'Tarea no encontrada'}), 404
+
+        # validamos si la tarea le pertenece al usuario autenticado
+        if tarea.usuario_id != usuario_id:
+            return jsonify({'ok': False, 'message': 'No tienes permisos'}), 403
         return jsonify({'ok': True, 'data': tarea.to_dict()})
     except Exception as e:
         return jsonify({'ok': False, 'message': str(e)}), 500
@@ -37,23 +43,22 @@ def obtener_tarea(id):
 
 # Crear una nueva tarea
 @tareas_bp.route('/', methods=['POST'])
+@jwt_required()
 def crear_tarea():
     try:
         payload = request.get_json()
+        usuario_id = int(get_jwt_identity())
 
         # validacion titulo
         if not payload.get('titulo'):
             return jsonify({'ok': False, 'message': 'El titulo es requerido'}), 400
-
-        if not payload.get('usuario_id'):
-            return jsonify({'ok': False, 'message': 'La tarea debe estar asociada a un usuario'}), 400
 
         # Guardar un registro en la base de datos
         nueva_tarea = Tarea(
             titulo=payload.get('titulo'),
             descripcion=payload.get('descripcion'),
             categoria=payload.get('categoria', ''),
-            usuario_id=payload.get('usuario_id')
+            usuario_id=usuario_id
         )
         db.session.add(nueva_tarea)
         db.session.commit()
@@ -64,13 +69,18 @@ def crear_tarea():
 
 
 @tareas_bp.route('/<int:id>', methods=['PUT'])
+@jwt_required()
 def actualizar_tarea(id):
     try:
         payload = request.get_json()
+        usuario_id = int(get_jwt_identity())
         tarea = Tarea.query.get(id)
 
         if tarea is None:
             return jsonify({'ok': False, 'message': 'Tarea no encontrada'}), 404
+
+        if tarea.usuario_id != usuario_id:
+            return jsonify({'ok': False, 'message': 'No tienes permisos'}), 403
 
         if 'titulo' in payload:
             tarea.titulo = payload.get('titulo')
@@ -85,11 +95,16 @@ def actualizar_tarea(id):
 
 
 @tareas_bp.route('/<int:id>', methods=['DELETE'])
+@jwt_required()
 def eliminar_tarea(id):
     try:
         tarea = Tarea.query.get(id)
+        usuario_id = int(get_jwt_identity())
         if tarea is None:
             return jsonify({'ok': False, 'message': 'Tarea no encontrado'}), 404
+
+        if tarea.usuario_id != usuario_id:
+            return jsonify({'ok': False, 'message': 'No tienes permisos'}), 403
 
         db.session.delete(tarea)
         db.session.commit()
